Finish impl of collection-type constants.
There's a lot of redundancy here wrt mapping ThriftType->trueType->TypeNames;
it could be streamlined.  Let's do that later, if we need to.

At this point it all works.  The only thing that irks me is that, for collections
in static blocks, we have to separate temporary declaration from initialization.
It's dictated by the interplay between `generateFieldInitializer` and
`renderConstValue`, but is displeasing.  Other than that, #2 is fixed.