Require references to included types to be scoped
Types in included files are exposed to including files only as scoped names -
that is, their names are prefixed with the filename (without extension).  For example:

```thrift
include 'base.thrift'

struct Foo {
  1: base.SomeType bar
}
```

Prior to this diff, included types were simply added to an ambient set of visible names;
this is contrary to the Thrift spec.  Oops!

Fixes #7.