Add TypeElement to the parser model, give it annotations.
In order to support annotations on types, we need to move beyond just Strings.
TypeElement models typenames in Thrift IDL, with different shapes to account
for the different kinds of type - scalar, list, set, and map.  All support
annotations.

This also opens up some nice simplifications in Linker and in Constant, but
that's a task for another day.