Remove Okio usage from transports
Instead of providing BufferedSources and BufferedSinks, Transports
now expose more traditional methods for reading and writing byte
arrays.  This change cascades through the various transport
implementations up to the Protocols that use them.

The motivation for this change is Okio's global SegmentPool behavior.
It is not safe to interact with a Buffer from different threads, e.g.
reading on one thread and closing from another.  This is a common
scenario, and it doesn't make sense to let our convenience restrict it.

Note that generated `binary` fields are still ByteString; this change
impacts Transports only.