Emit const references where appropriate.
ConstantBuilder wasn't aware of refernces to other consts when
rendering; this would result in anything from a compiler crash to
incorrect values being emitted.  This is fixed here.

This is definitely more duct-tape than architecture; I punted on a good
solution for ensuring that qualified names are correctly resolved.  We
should think about how to represent this kind of scoping in the final
linked Schema - that, or teach fields to be aware of this sort of
indirection and validate properly during linking.

Fixes #35.