Allow constants and types with same name (#59)
This practice, while of dubious merit, is legal in Thrift; this commit
adds support.

This took a surprising amount of effort; before, we have always assumed
that a single name would map to at most one `Named` instance - now,
that is no longer the case.  `Linker` correspondingly has grown a new
lookup method, and things that use it (i.e. `Constant` validation) now
need to ensure that they use the correct lookup method.