Rewrite parser with ANTLR (#104)
Time for another Very Large Change! Here, finally, is the end of the old hand-rolled parser - and, hopefully, the end of fiddly little bugs around trailing-doc comment parsing.

This change should be nearly transparent to consumers; the only visible difference is in parser error reporting. Previously, we would just throw an exception at the first encountered error. We still throw an exception, but now gather all encountered errors into a message.

This is in anticipation of refactoring error reporting a bit; for now, I'd prefer to keep a stable API.

In the process of implementing this, I upgraded AutoValue and finally got tired of the Gradle plugin we were using - I've replaced it with tbroyer's apt plugin, which is comparatively well-behaved.

Fixes #30.