Add UUID to parsing for matching elements across instances (#152)
* Add UUID to parsing for matching elements across instances

This is a prototype of trying to tag parsed elements with unique IDs so that one could match them even if you `newBuilder`'d it, you could match back other elements that reference the now-old reference to them. This would allow for one to normalize references in a `Schema` after changing a bunch of elements in it, such as with a preprocessor to edit Schemas before code gen.

If the implementation here looks good, I can proceed with figuring out how to update tests to handle this in a deterministic way

* Create ThriftyParserPlugins for hooks into UUID generation

This makes it more testable as well as flexible if others have their own needs. This could also be used to conceivably turn it off, if they don't care about it and want the slight perf gain

* Defer to ThriftParserPlugins for UUID generation in elements

* Update ThriftParserTest to use plugins