Port non-ThriftType schema types to Kotlin (#179)
Linker, Schema, Loader, Program, and friends are herein converted to Kotlin. A few methods that should by rights be internal are left public here due to the fact that Java unit-tests cannot see the internal members, even though they are part of the same "module". We'll hide them again once the ThriftType family is also ported.

ThriftType and descendants are tightly intertwined, and it will be tricky to disentangle them. I'm vaguely hopeful that a sealed-class hierarchy will fall out of the mess, but the way forward is still unclear.

Finally, there are a plethora of @JvmName annotations on properties here, to avoid too many spurious changes in unrelated files. We'll get rid of them all after ThriftType is in kt.