Port EnumType, UserElement, StructType to Kotlin
This has exposed an unfortunate design mistake from early in the
project.  Schema types are invalid when constructed, and are only valid
after linking.  We've hacked around this fact in unit tests with
mocking, but idiomatic Kotlin is _not_ mockable.  I don't think we
should contort ourselves to keep using mocks, but pulling on the thread
in ConstantTest revealed that we basically need to run the entire
parse/validate/link chain to do some small unit tests.  This is
certainly a sore spot, which we're working around here with reflection.

Longer-term, we'll probably end up splitting the types into unlinked and
linked variants, and only publicly exposing the linked types.  For now,
reflection suffices.