Port ThriftType to Kotlin :)
This is the final piece of thrifty-schema to be ported, and it touches
*lots* of stuff.  I've redefined some of the abstract funs as vals,
because it would have been as large of a change to make them work as
funs, so why not do it right the first time.  In the process,
UserElementMixin has become a data class - the first of many to come!