Port the remainder of thrifty-schema to Kotlin (#180)
This PR represents a series of smaller commits, each one porting a piece of the ThriftType hierarchy, roughly from the leaves down to the root, culminating with the port of ThriftType itself. It's a massive change - we will have to trust the integration tests to confirm that things are still working as intended.

This does not represent the end of our efforts - we can make much more idiomatic use of Kotlin than this. In particular, the porting effort has thrown a spotlight on a latent design flaw - the UserTypes are invalid when created, which makes fine-grained unit testing extraordinarily difficult without reflection or mocks. We may decide to live with it, or we might also decide to split up ThriftType into pre- and post-linking immutable representations.

Finally, we haven't begun to explore using sealed classes or data classes, which promise to eliminate a good amount of boilerplate and infrastructure code.