Change command-line flags
Instead of `--emit-kotlin`, we're adding a `--lang` option defaulting to
Java.  Instead of `-use-java-style-names`, we're adding `--name-style`
defaulting to, well, "default", but that can also be "java" (or
potentially other styles as well).

This change also attempts to infer language if it is unspecified - in
this case, if `--kt-file-per-type` is specified, but `--lang` is not, we
assume that `--lang=kotlin` was intended.