Remove "Generated from:" from constants' JavaDoc
This line unfortunately includes the local file system path that can
differ between individual systems. This isn't ideal for projects that
store these generated files in source control because is causes
unnecessary file differences.

This could also be controlled by omitFileComment, but that flag is
currently documented as only controlling the top-level header comment.
Redefining it and using it here would be a good alternate solution
because projects that want that behavior will also want this one.