Canonicalize paths of all loaded thrift files
This will fix the issue observed on macOS wherein the same file can be
loaded multiple times, if its canonical path is not the same as its
normalized path and it is included by another loaded file.

I'm dying to know if this has ever been a problem outside of unit tests
on macOS - hopefully not - but after this, it won't be.