Specify that a successful parse consumes all input
Prior to this commit, our generated antlr parser would happily stop
midway through a file if it discovered a token it couldn't handle, such
as a semicolon at the top level.  This was treated as a successful
parse, leading to surprising cases of missing generated structs.

Fixed here by adding `EOF` to the end of our top-level `document`
definition, which is the antlr way of saying "consume all input or else
it is an error".

Fixes #297