Implement default values for unions (#308)
This adds support in generated Kotlin code for unions that have default values.

Unions are currently represented as sealed types; the default value, if present, will be a static val on the base sealed type. For example, given the following thrift:

```
union Sample {
  1: string Foo = "bar";
}
```

this will emit (partial code):

```
sealed class Sample : Struct {
  companion object {
    @JvmStatic
    val DEFAULT: Sample = Foo("bar")
  }

  ...
  class Builder {
    private var value: Sample? = DEFAULT
   ...
  }

  ...
  class Adapter {
    override fun read(protocol: Protocol): Sample {
      var result: Sample = DEFAULT
      ...
    }
  }
}
```

The only point about which I'm uncertain is the Adapter.read method; it's not clear to me whether we should assume a default value if no value (or no expected value) is received; the remote end may have a different idea about what the default value is. I expect we'll never revisit this point due to the obscurity of the feature, but have called out the question as a TODO.

Fixes #303.