Introduce an --omit-service-clients option (#330)
When specified, no service clients will be generated. We only support
service clients in the Kotlin generator, so this is implicitly a
Kotlin-only option, but it's not prefixed with `kt` because it could
apply to Java in the future.

This is useful for projects that only need Thrift-generated structures,
etc. and won't use Thrift service calls.