Introduce a --kt-emit-jvmname option
When specified, `@file:JvmName(name)` annotations will be added to all
of the generated files. This improves the usability of importing these
Kotlin-based classes into Java code modules (e.g. as `foo.Constants`
instead of `foo.ConstantsKt`).

This would ideally be the default behavior, but it's introduced as an
opt-in compiler option to avoid breaking backwards compatibility.