Explicitly call fail() on a failed request when an exception occurs
If an exception occurs when performing a client request (i.e. an
IOException or a RuntimeException) then the request never completes,
even with a fail.

This happens because these types of exceptions are handled differently
in `AsyncClientBase.invokeRequest()` than other types, and the
exception is re-thrown without `fail`-ing the request leaving
client code hung.

This patch fixes the problem #358.

Signed-off-by: Alexey Morozov <morozov@gmail.com>